requirements.txt
fastapi
uvicorn
requests
python-dotenv
agno
playwright
pandas
beautifulsoup4
lxml
openai
SpeechRecognition
pydub
